+++ src/jp/crestmuse/cmx/elements/BaseDynamicsEvent.java
@@ -4,7 +4,7 @@ public class BaseDynamicsEvent extends MutableMusicEvent {
 
   private double value;
   
-  public BaseDynamicsEvent(int timeStamp, double value, int ticksPerBeat){
+  public BaseDynamicsEvent(long timeStamp, double value, int ticksPerBeat){
 //  public BaseDynamicsEvent(int timeStamp, int ticksPerBeat, double value){
   super(Type.CONTROL_CHANGE, timeStamp, timeStamp, ticksPerBeat);
     this.value = value;
